Implementation of import only by playlist feature, various improvements to structure and function:
- Updated Qx (support for generic QString join function)
- Added Playlist Game Mode option, "Selected Platforms Only" replicates
existing functionality, "Force All" allows importing every game on a
playlist regardless of platform and performing an import with only
playlists selected
- Renamed the FP::Install query functions to be more generic and
descriptive of their purpose since some are now used multiple times and
aren't necessarily tied to ImportWorker
- Added ability to filter queryGamesByPlatform() results by game ID
(required for playlist Force All function)
- FP::Install query functions that return lists now only append a result
if its size is > 0 so that the content of the results are more
immediately apparent
- Added FP::Install::queryPlaylistGameIDs() for getting a list of game
IDs that belong to the provided playlists (required for playlist Force
All function)
- DBQueryBuffer::size is now explicitly initialized to 0 to prevent
unintended behavior when checking the size of a freshly initialized
instance
- Corrected several typos of old class names in comment headers within
cpps
- Broke ImportWorker::doImport() down into several reusable, step-based
private functions for more clear logic flow
- Changed some working variables in ImportWorker to members in order to
facilitate the above
- Clarified plurality of some database query results in ImportWorker
- Replaced all logic for enabling/disabling UI elements with a call to
refreshWidgetEnableStates(), which utilizes a hash-map of boolean return
lambdas to determine which elements should be in either state
- Simplified the calls to reset() for both install pointers by moving
them from failure conditions to before each time the install are checked
for validity since the previous install link (if present) will be broken
by the end of the check either way. This also ensures the old instance
is always dropped, as this action was missing in a few places
previously, which is no doubly important since some widgets check these
pointers as part of their enable condition
- Created function clearListWidgets() instead of requiring individual
calls to clearing both widgets since they are always cleared together
- Removed customSetUpdateGroupEnabled() and now instead directly control
the enable state of ui->groupBox_updateMode. The disabled font color of
the "Existing Item" label has been altered to be legible, removing the
need for the old method
- Renamed getSelectedImageOption() to getSelectedImageMode() for
consistency
- Selecting existing playlists now also enables the Update Mode group
box since they are affected by the "Remove Obsolete" option
independently of when existing platforms are selected (which previously
was thought as the only situation when playlists were affected by the
options, making enabled the box from playlists alone unnecessary)
- The playlist game Force All option enables the Update Mode group box
because its possible that existing platforms will be affected by the
import when used in with that option selected
- Changed warning of importing with selected platforms selected to be
more generic so that it could be deployed when existing playlists or the
playlist game Force All mode is used as well
- Corrected several typos in runtime_error throw statements when
concerting sender() pointers in MainWindow slots
- mAlteringListWidget guard was removed and replaced with hash maps that
track the check state of the items within the widgets so that when
all_on_listWidget_itemChanged() is called it can be determined if the
call was due to a change in check state, making the guard unnecessary
- Replaced some bolding in help messages with italics to reserve bold
for the option names only
- Added convenience functions for checking if an existing platform or
playlist is selected