Build CLIFp from source instead of requiring an externally supplied copy
FetchContent adds CLIFp as a subproject that is build, a custom command
and target are then used to copy the executable into a resource folder
within OFILb's build directory at build time. The function
add_generated_resources_collection() is used to create a resources.qrc
file in this build-based resource folder which is then added to the
OFILb executable target.

add_dependencies() is used to ensure this file-copy target is "built"
before OFILb is.

Essentially there is now a seperate resource location that is
"dynamically" generated and merged with the static one in the source
folder.