Fix saving of existing data documents
Previously existing documents were opened with OpenMode ReadOnly, but
not closed. Fe::Install::saveDataDocument would then try to open the
already open file with open mode WriteOnly, which would do nothing, and
then the document would fail to write since the open mode was still
ReadOnly.

The straightforward fix would have been to make sure the underlying file
on disk was closed after calling openDataDocument, but it seemed more
appropriate to keep a handle to that file while the doc is "open" in the
context of the importer.

So instead, the file was changed to be opened in ReadWrite mode, with
the open attempt in saveDataDocument being removed.

Additionally, the resize() call on the file when it comes time to save
it is now checked for success in lieu of the now removed open attempt
there.