Move file management responsibility to readers/writers
The disadvantage of this is that file handles can no longer be kept open while the DataDoc is open in ImportWorker;
however, it allows for writers/readers to be implemented that rely on classes that must manage the file handle
themselves (i.e. QSqlDatabase with an SQLite database), and is also a step towards supporting potential data docs
that aren't file based at all (as unlikely as those may be).