Overhaul Install interface
- Introduce InstallFoundation: As much of the core Install logic that
should be the same for all Install derivatives is moved to its this new
base class that the actual Install interface inherits. This way the
aspects of Install that are relevant when writing new derivatives are
clearer and the composition of the class is less cluttered. Generally
such derivatives don't need to concern themselves with InstallFoundation
at all, but there are still a few small exceptions

- Removed the "-derived()" functions of Install in favor of the approach
in which the parent function is overridden and then called within the
child function. This is technically less safe as before the base
behavior being executed was guaranteed and now it could be forgotten,
but this heavily reduces clutter and as long as each new install is made
while referencing an existing one it should be clear when any base
implementation needs to be invoked

- Rework image handling: While to somedegree these changes resulted in a
shuffling of implementation between ImportWorker and Install, the
overall image import process is simplified and less redundant. Instead
of the various functions used to cover all potential manners in which an
Install could reference images, the install is now provided with the
selected ImageMode and handed image info in two distinct places, itself
being entirely responsible for how to use said info. Per game image
paths are passed through the currently opened platform, which covers if
the install uses in-platform images. Additionally, the base
implementation of Fe::PlatformDoc ensures that these paths are passed
through to the parent Install for ones that reference individual games
in separate documents, or so they can determine the destination path of
those images in the case of a Copy/Link. The notifier
preImageProcessing(QList<ImageMap>& workerTransfers, ImageSources
bulkSources) is used to to complete the handling of this case, as well
as when an Install references images in bulk. The argument `bulkSources`
provides the FP image folder paths, while `workerTransfers` is to be
filled with the previously mapped (sourcePath->destinationPath) images
in the Copy/Link case so that ImportWorker can handle the transfers.
Generally it is good to store the mapped images in a list and then swap
the contents of that list with this argument list for an almost zero
cost exchange.

- Because of the above change, PlatformDoc::addGame() no longer needs to
return a pointer of the converted game, significantly simplifying error
handling of the doc and allows for the potential of returning an error
object for immediate handling instead of void in the future as an
alternative to the current system where a doc can have an error status
that has to be checked for

- Installs now populate their existing doc set with all docs that belong
to them, not just ones relevant to Flashpoint. Mainwindow now handles
filtering to only the relevant ones when checking for if an update is
going to be performed.

- Installs now provide translateDocName() (with a default implementation
that does nothing in case an install doesn't need it) to ensure FP doc
names are translated to meet their restrictions before being used
through the rest of the interface

- Changed open/save to checkout/commit (might change to lease/return or
something else later) since calling those functions may not actually
correspond directly to a file being opened/closed on disk depending on
an install's implementation

- Updated libfp so that its item getters are property style

- Added MainWindow::selectionsMayModify() for simplified checking of
when the user's import selections may cause modifications to exist
Platforms/Playlists