Use Fp::Set to group games with their add apps for import
By using a QMultiHash to cache additional apps by parent ID, the
complexity of adding the correct additional apps to a given playlist is
greatly reduced since they can now be grabbed directly instead of of the
cache needing to be scanned by every platform.

This also allows for frontends to have the parent info of every
additional app readily available if required.

This has somewhat complicated the Fe::DataDoc situation given that Sets
can't realsitically be an updateable item since theyre just a container
(and the update logic must be performed on the games/addapps directly),
but for now its OK and can be worked on later.