Change Am::Romlist to inherit from UpdateableDoc with type ConfigDoc
Inheriting from Fe::PlatformDoc worked and was more straightforward, but
it wasn't fully accurate. In the context of the interface,
Am::PlatformInterface is its true "PlatformDoc" and Romlist acts more
like a config doc, especially since it is only used internally by
Am::Install and not touched directly by ImportWorker.

The old inheritance also created the issue that when the romlist was
modified it would be added to the modified doc ledger as a PlatformDoc,
which means that Fe::Install::modifiedPlatforms() would include it. This
caused its use to fumble since the idea of the function is to only
return modified platforms in a strict interface/ImportWorker context
without including anything internal to a given install.