Properly handle Parents.xml
The previous approach did not handle additional, unseen fields that can
be present within a Parents item. This approach is a bit less performant
due to the use of a list, but there isn't a great way to collect Parent
entries in a map since they have no truly unique way to identify them.