Set reasonable max for LB free index tracker
Previous max of -1 was a leftover from the previous implementation of
Qx::FreeIndexTracker who's index type was templated, so -1 represented a
reasonable max (likely 32-bit integer limit); however, now that the
tracker uses quint64 for indexing, -1 was interpreted as the maximum
value for 64-bit integers and caused allocaiton of its internal bit-
array to fail due to absurd size requirements.

This was undiscovered due to the previous playlist bugs.