LB: Use modified playlist ID cache directly
No need to store as a QHash with the playlist name as the key since the
name is not needed. This also avoids the awkwardness that is the
conversion between the original playlist name and the translated name,
specific to LB. Before, playlists with names that differed between their
original and translated names were not being added to the Flashpoint
platform category as the keys in the hash were the pre-translated names,
while the list checked against at the end were the translated names.