Distinguish between modified docs and deleted docs
Frontends often use modifiedPlatforms()/modifiedPlaylists() to check
which Platforms/Playlists were affected by the important and need to be
taken into account by other parts of the frontend; however, this
obviously only applies if the doc was stricly modified and not outright
removed, which can happen if the commited modified doc is empty.

This change ensures that only modified docs that remain will be part of
the modified list.