Fix filenames changing when running LB after import
LaunchBox seems to have new rules for updating the names of files
in the data directory, in particular Playlists, in which it attempts
to change the filename to match the <Name> attribute name and then
apply those rules to deal with illegal characters.

This was causing icon names to be mismatched, meaning they wouldn't
show in LaunchBox, and playlist names to not match on future imports,
such that FIL didn't know they were the same playlist.

Better kosherize filenames to try and match what LB would do so that
it doesn't change them after FIL has touched them.