Add templated interum classes for easier Launcher implementation
Avoids the need for manual static_casts in launcher install classes and
also allows for some further automation for the same.

This allows for the install registry to be part of a template
traits system and reduce the degree to which a macro is needed for
registration.

Also, move the responsibility of image task storage entirely to Worker,
and have IPlatformDoc implement a sensible default for addSet() which
forwards image processing to Install if the right mode is selected.

Move item related concepts to lr-items-interface.h and use inline
constraints.