Consolidate update facilities into a flexible container type
Adds IUpdatableDoc::UpdatableContainer which provide various methods
that replace the use of the standalone update functions that required
manually passing containers and keys.

Internally this uses a std::unordered_set instead of a QHash which saves
space. Instead, items are setup to be hashed/compared using only the
portion of them that makes them uniquely identifiable (i.e. what used to
be used as the hash key).

Also:
- Corrects not passing QUuid by const& where sensible
- Adds Builder base for use with types that need one but do not derive
from Item