Add support for ES-DE
Still could use some work, but has all of the essentials.

There is an annoying requirement to manage dummy files in order to
satisfy the ROM scanner, les the games not appear despite there being
the metadata for them. There is an option that disables this reliance
but its global and would be impossing to require that users enable it.
Eventually I'd like to make a PR for ES-DE that allows a config settings
on a system-by-system basis to enable this behavior, but they don't seem
very receptive to PRs.