Always ensure import properties are refreshed when installs change
If an install path was changed to another valid install path, because
the corresponding property was never set to nullptr in the interim, the
matching "install ready" proprerty would simply remain "true" meaning
that the mBothInstallsReady property would not toggler and therefore not
fire its notifier. This would result in the new install not being
checked for it's existing docs and ultimately a loop where the "installs
changed" check would refresh the docs, making it look like they changed
compared to the initially empty list. But then, when that would force a
re-eval of the installs, that would again not fire the doc check and
then the process would repeat.

Now we ensure that both install types are reset to nullptr before
checking a new path, which will ensure that dependent properties change
if the pointer originally contained an install, and therefore change
again when the new install (if valid) is loaded.